+++ R/app_server.R
@@ -92,4 +92,6 @@ app_server <- function(input, output, session) {
   mod_ai_annotation_server("ai_annotation_1", global_data = global_data, prj_init = prj_init)
   # PlantCyc database construction toolkit
   mod_plantcyc_database_server("plantcyc_database_1")
+  # KEGG organism database construction toolkit
+  mod_kegg_database_server("kegg_database_1")
 }